misc: remove Makefile and fix warning

+++ src/cpu/gol-cpu.cc
@@ -3,13 +3,13 @@
 void gol_cpu(int width, int height, int n_iterations)
 {
     Map m(height, width);
-    for (size_t i = 0; i < n_iterations; i++)
+    for (int i = 0; i < n_iterations; i++)
         m.basic_cpu_compute();
 }
 
 void gol_cpu_parallel(int width, int height, int n_iterations)
 {
     Map m(height, width);
-    for (size_t i = 0; i < n_iterations; i++)
+    for (int i = 0; i < n_iterations; i++)
         m.parallel_cpu_compute();
 }
